# docs/openapi.yaml
basePath: /api
definitions:
...
  handlers.CreateTeamRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  models.ScoreboardEntry:
    properties:
      last_solve:
        type: string
      points:
        type: integer
      rank:
        type: integer
      solve_count:
        type: integer
      team_id:
        type: string
      team_name:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  models.Team:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      invite_id:
        type: string
      invite_permission:
        type: string
      name:
        type: string
      owner_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8090
info:
  contact: {}
  description: Self-hosted CTF platform API. Most write endpoints require authentication
    via JWT cookie (auth_token).
  title: hCTF API
  version: 1.0.0
paths:
...
  /teams:
    get:
      description: Invite codes are only visible to team members.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List all teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Team data
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a new team
      tags:
      - Teams
  /teams/{id}:
    get:
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              members:
                items:
                  type: object
                type: array
              team:
                $ref: '#/definitions/models.Team'
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get a team with its members
      tags:
      - Teams
  /teams/disband:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Disband the team (owner only)
      tags:
      - Teams
  /teams/invite-permission:
    post:
      consumes:
      - application/json
      parameters:
      - description: 'Permission value: ''owner_only'' or ''all_members'''
        in: body
        name: body
        required: true
        schema:
          properties:
            permission:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update who can see the team invite code (owner only)
      tags:
      - Teams
  /teams/join/{invite_id}:
    post:
      parameters:
      - description: Team invite code
        in: path
        name: invite_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Join a team using an invite code
      tags:
      - Teams
  /teams/leave:
    post:
      description: Team owners must transfer ownership or disband before leaving.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Leave the current team
      tags:
      - Teams
  /teams/regenerate-invite:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              invite_id:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Regenerate the team invite code (owner only)
      tags:
      - Teams
  /teams/scoreboard:
    get:
      description: Returns HTML table when called from HTMX (HX-Request header), otherwise
        returns JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ScoreboardEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get team scoreboard rankings
      tags:
      - Teams
  /teams/transfer-ownership:
    post:
      consumes:
      - application/json
      parameters:
      - description: New owner user ID
        in: body
        name: body
        required: true
        schema:
          properties:
            new_owner_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Transfer team ownership to another member (owner only)
      tags:
      - Teams
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"